Quiz 4 and homework 2 completed

+++ homework-02/bugs/12fpcount.cc
@@ -10,8 +10,8 @@
 
 // Corrected:
 int main() {
- 	float int sum = 0;
-  	for (float int i = 1; i <= 100000000.0f; i++)
+ 	uint64_t sum = 0;
+  	for (uint64_t i = 1; i <= 100000000; i++)
   		sum += i;
-  	return sum;
+	return sum;
 }
